requirements.txt
Pillow ~= 9.4.0
discord.py ~= 2.2.3
PyNaCl ~= 1.5.0